Correcting how to get keys from a hash type.

+++ src/client.py
@@ -17,11 +17,11 @@ def gol_connection():
 # queue
 def read_from_redis(key):
     with gol_connection() as connection:
-        record = connection.get(key)
-        print record
+        dimension = connection.hget(key, 'dimensions')
+        message = connection.hget(key, 'message')  # nothing returned from this yet.
 
-        x, y = tuple(float(num) for num in record['dimension'].split(','))
-        array = numpy.fromstring(record['message'], dtype='<i4').reshape((x, y))
+        x, y = tuple(float(num) for num in dimension.split(','))
+        array = numpy.fromstring(message, dtype='<i4').reshape((x, y))
         return array
 
 
@@ -31,13 +31,16 @@ def write_to_redis(key, array):
             'dimension': "%s,%s" % array.shape,
             'message': array.tostring('F')  # fortran order
         }
-        connection.set(key, data)
+        connection.lpush(key, data)
 
 
 # Read from the Game Of Life key/value pair.
 def read_gol():
     with gol_connection() as connection:
-        loaded_array = numpy.fromstring(connection.brpop('A'), dtype='<i4')
-        array = loaded_array.reshape((100, 100))
-        pyplot.imshow(array)
-        pyplot.savefig('scene.png')
+        if connection.exists('A'):
+            loaded_array = numpy.fromstring(connection.brpop('A'), dtype='<i4')
+            array = loaded_array.reshape((100, 100))
+            pyplot.imshow(array)
+            pyplot.savefig('scene.png')
+        else:
+            print 'No key `A`.'
